Add config options related to Vi input mode
Setting `vi_start_in_nav_mode` to `True` enables `NAVIGATION` mode on startup.
The issue is that due to the current behaviour of `ViState.reset()` input mode
gets resetted back to `INSERT` on the every iteration of the main loop. In order
to at one hand to provide the user with desired behaviour and on the other hand
doesn't introduce breaking changes the other option `vi_keep_last_used_mode` was
introduced which sets `input_mode` to the state observed before reset.

`vi_keep_last_used_mode` can be useful even with `vi_start_in_nav_mode` set to
`False` in the case the user prefer to start in `INSERT` mode but still wants to
maintain the last mode he was in.

In the case of `vi_keep_last_used_mode` set to `False` and
`vi_start_in_nav_mode` to `True` `NAVIGATION` mode is set on every iteration the
same way `INSERT` was set before this commit.

Fixes #258.

Commit rebased and modified by Jonathan Slenders.