Fix exit confirmation.
Before, the exit confirmation was not focused. Which meant that key bindings of
the main buffer were still active. If we are in Vi mode, that meant that there
was a key binding for the ("y", "y") already, which caused the handling of "y"
to be delayed (it was not marked as eager). This fix will focus the exit
confirmation and avoid further interference of buffer key bindings.