Improve dictionary completion performance.
This improves the performance for dictionary-like objects where iterating over
the keys is fast, but doing a lookup for the values is slow. This change
ensures we only do value lookups when really needed.

The change also caches the meta text so that we don't have to recompute it
during navigation of the completion menu.